Halfway through moving lots of stuff into BleEndpoint/Config
