+++ src/com/idevicesinc/sweetblue/BleDeviceConfig.java
@@ -2,7 +2,6 @@ package com.idevicesinc.sweetblue;
 
 import java.util.UUID;
 
-import android.bluetooth.BluetoothDevice;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Build;
@@ -16,7 +15,6 @@ import com.idevicesinc.sweetblue.BleManager.DiscoveryListener.LifeCycle;
 import com.idevicesinc.sweetblue.annotations.Nullable;
 import com.idevicesinc.sweetblue.annotations.Immutable;
 import com.idevicesinc.sweetblue.annotations.Nullable.Prevalence;
-import com.idevicesinc.sweetblue.utils.State;
 import com.idevicesinc.sweetblue.utils.*;
 
 /**
@@ -33,7 +31,7 @@ import com.idevicesinc.sweetblue.utils.*;
  * <br><br>
  * TIP: You can use {@link #newNulled()} (or {@link #nullOut()}) then only set the few options you want for {@link BleDevice#setConfig(BleDeviceConfig)}.
  */
-public class BleDeviceConfig extends BleServerConfig implements Cloneable
+public class BleDeviceConfig extends BleEndpointConfig implements Cloneable
 {
 	public static final double DEFAULT_MINIMUM_SCAN_TIME				= 5.0;
 	public static final int DEFAULT_RUNNING_AVERAGE_N					= 10;
@@ -328,28 +326,6 @@ public class BleDeviceConfig extends BleServerConfig implements Cloneable
 	 */
 	@Nullable(Prevalence.NORMAL)
 	public ReconnectPersistFilter reconnectPersistFilter_shortTerm			= new DefaultReconnectPersistFilter(DefaultReconnectPersistFilter.SHORT_TERM_TIMEOUT);
-
-	/**
-	 * Default is an instance of {@link com.idevicesinc.sweetblue.BleDeviceConfig.DefaultHistoricalDataLogFilter} -
-	 * set an implementation here to control how/if data is logged.
-	 */
-	@com.idevicesinc.sweetblue.annotations.Advanced
-	@Nullable(Prevalence.NORMAL)
-	public HistoricalDataLogFilter historicalDataLogFilter					= new DefaultHistoricalDataLogFilter();
-
-	/**
-	 * Implement this to override the default behavior, which is simply to return an instance created with
-	 * the constructor {@link HistoricalData#HistoricalData(byte[], com.idevicesinc.sweetblue.utils.EpochTime)}.
-	 */
-	@com.idevicesinc.sweetblue.annotations.Advanced
-	@Nullable(Prevalence.NORMAL)
-	public HistoricalDataFactory historicalDataFactory						= new HistoricalDataFactory()
-	{
-		@Override public HistoricalData newHistoricalData(final byte[] data, final EpochTime epochTime)
-		{
-			return new HistoricalData(data, epochTime);
-		}
-	};
 
 	/**
 	 * As of now there are two main default uses for this class...
@@ -921,314 +897,6 @@ public class BleDeviceConfig extends BleServerConfig implements Cloneable
 				return Please.persistIf(e.totalTimeReconnecting().lt(m_timeout));
 			}
 		}
-	}
-
-	/**
-	 * Provide an instance of this class to {@link com.idevicesinc.sweetblue.BleDeviceConfig#historicalDataLogFilter} to control
-	 * how/if historical data from BLE operations is logged.
-	 */
-	@com.idevicesinc.sweetblue.annotations.Advanced
-	@com.idevicesinc.sweetblue.annotations.Lambda
-	public static interface HistoricalDataLogFilter
-	{
-		/**
-		 * Special value you can use in place of Java's built-in <code>null</code>, just for code readability.
-		 */
-		public static HistoricalDataLogFilter DISABLED = null;
-
-		/**
-		 * Signifies where the data came from, usually from a BLE read or notification.
-		 */
-		public static enum Source implements UsesCustomNull
-		{
-			/**
-			 * Satisfies soft contract of {@link com.idevicesinc.sweetblue.utils.UsesCustomNull}
-			 */
-			NULL,
-
-			/**
-			 * Originates from {@link BleDevice#read(java.util.UUID, BleDevice.ReadWriteListener)}.
-			 *
-			 * @see com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Type#READ
-			 */
-			READ,
-
-			/**
-			 * Originates from {@link BleDevice#startPoll(java.util.UUID, com.idevicesinc.sweetblue.utils.Interval, BleDevice.ReadWriteListener)}.
-			 *
-			 * @see com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Type#POLL
-			 */
-			POLL,
-
-			/**
-			 * Originates from {@link BleDevice#enableNotify(java.util.UUID, com.idevicesinc.sweetblue.BleDevice.ReadWriteListener)}.
-			 *
-			 * @see com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Type#NOTIFICATION
-			 */
-			NOTIFICATION,
-
-			/**
-			 * Originates from {@link com.idevicesinc.sweetblue.BleDevice#enableNotify(java.util.UUID, com.idevicesinc.sweetblue.BleDevice.ReadWriteListener)}.
-			 *
-			 * @see com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Type#INDICATION
-			 */
-			INDICATION,
-
-			/**
-			 * Originates from {@link com.idevicesinc.sweetblue.BleDevice#enableNotify(java.util.UUID, Interval, com.idevicesinc.sweetblue.BleDevice.ReadWriteListener)},
-			 * where a force-read timeout is invoked, or from {@link BleDevice#startChangeTrackingPoll(java.util.UUID, com.idevicesinc.sweetblue.utils.Interval, BleDevice.ReadWriteListener)}.
-			 *
-			 * @see com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Type#PSUEDO_NOTIFICATION
-			 */
-			PSUEDO_NOTIFICATION,
-
-			/**
-			 * Originates from manual addition of data through {@link BleDevice#addHistoricalData(UUID, byte[], EpochTime)}
-			 * or various overloads that add one piece of historical data.
-			 */
-			SINGLE_MANUAL_ADDITION,
-
-			/**
-			 * Originates from manual addition of data through {@link BleDevice#addHistoricalData(UUID, java.util.Iterator)}
-			 * or various overloads that add multiple pieces of data all at once.
-			 */
-			MULTIPLE_MANUAL_ADDITIONS;
-
-			/**
-			 * Returns <code>true</code> if <code>this</code> == {@link #NULL}.
-			 */
-			@Override public boolean isNull()
-			{
-				return this == NULL;
-			}
-		}
-
-		/**
-		 * Event passed to {@link BleDeviceConfig.HistoricalDataLogFilter#onEvent(HistoricalDataLogEvent)} that provides
-		 * information you can use to determine whether or not {@link HistoricalDataLogEvent#data()} should be logged.
-		 */
-		public static class HistoricalDataLogEvent
-		{
-			/**
-			 * The device in question.
-			 */
-			public BleDevice device()  {  return m_device;  }
-			private final BleDevice m_device;
-
-			/**
-			 * The data to be written.
-			 */
-			public byte[] data()  {  return m_data;  }
-			private final byte[] m_data;
-
-			/**
-			 * Timestamp of when the data was obtained.
-			 */
-			public EpochTime epochTime()  {  return m_epochTime;  }
-			private final EpochTime m_epochTime;
-
-			/**
-			 * The source of the data - read, notify, etc.
-			 */
-			public Source source()  {  return m_source;  }
-			private final Source m_source;
-
-			/**
-			 * The characteristic {@link java.util.UUID} associated with {@link #data()}.
-			 */
-			public UUID charUuid()  {  return m_charUuid;  }
-			private final UUID m_charUuid;
-
-			HistoricalDataLogEvent(final BleDevice device, final UUID charUuid, final byte[] data, final EpochTime epochTime, final Source source)
-			{
-				m_device = device;
-				m_charUuid = charUuid;
-				m_data = data;
-				m_epochTime = epochTime;
-				m_source = source;
-			}
-
-			/**
-			 * Returns true if this event is associated with the given uuid.
-			 */
-			public boolean isFor(final UUID uuid)
-			{
-				return uuid.equals(charUuid());
-			}
-
-			/**
-			 * Returns true if this event is associated with any of the given uuids.
-			 */
-			public boolean isFor(final UUID[] uuids)
-			{
-				return Utils.contains(uuids, charUuid());
-			}
-		}
-
-		static enum PersistenceLevel
-		{
-			NONE, MEMORY, DISK, BOTH;
-
-			public boolean includesMemory()
-			{
-				return this == MEMORY || this == BOTH;
-			}
-
-			public boolean includesDisk()
-			{
-				return this == DISK || this == BOTH;
-			}
-		}
-
-		/**
-		 * Special value returned from {@link BleDeviceConfig.HistoricalDataLogFilter#onEvent(HistoricalDataLogEvent)}
-		 * that determines if/how {@link HistoricalDataLogFilter.HistoricalDataLogEvent#data()} will get logged.
-		 */
-		public static class Please
-		{
-			final PersistenceLevel m_logChoice;
-
-			private byte[] m_amendedData = null;
-			private EpochTime m_amendedEpochTime = null;
-			private Long m_logLimit = null;
-
-			private Please(final PersistenceLevel logChoice)
-			{
-				m_logChoice = logChoice;
-			}
-
-			/**
-			 * Returns the limit provided through {@link #andLimitLogTo(long)}, or {@link Long#MAX_VALUE} if not applicable.
-			 */
-			public long getLimit()
-			{
-				return m_logLimit != null ? m_logLimit : Long.MAX_VALUE;
-			}
-
-			/**
-			 * Returns the amended data provided through {@link #andAmendData(byte[])}, or <code>null</code> if not applicable.
-			 */
-			@Nullable(Prevalence.NORMAL) public byte[] getAmendedData()
-			{
-				return m_amendedData;
-			}
-
-			/**
-			 * Returns the amended epoch time provided through {@link #andAmendEpochTime(EpochTime)}, or {@link com.idevicesinc.sweetblue.utils.EpochTime#NULL} if not applicable.
-			 */
-			@Nullable(Prevalence.NEVER) public EpochTime getAmendedEpochTime()
-			{
-				return m_amendedEpochTime != null ? m_amendedEpochTime : EpochTime.NULL;
-			}
-
-			/**
-			 * Last chance to amend or replace {@link HistoricalDataLogEvent#data()} before it's written to the log.
-			 *
-			 * @return <code>this</code> so you can chain calls together.
-			 */
-			public Please andAmendData(final byte[] data)
-			{
-				m_amendedData = data;
-
-				return this;
-			}
-
-			/**
-			 * Last chance to amend or replace {@link HistoricalDataLogEvent#epochTime()} before it's written to the log.
-			 * @return <code>this</code> so you can chain calls together.
-			 */
-			public Please andAmendEpochTime(final EpochTime epochTime)
-			{
-				m_amendedEpochTime = epochTime;
-
-				return this;
-			}
-
-			/**
-			 * Calling this will crop the log to the given limit <i>before></i> {@link HistoricalDataLogEvent#data()} is written.
-			 * So if you call this with <code>0</code> {@link BleDevice#getHistoricalDataCount(java.util.UUID)} will return <code>1</code>
-			 * after this.
-			 *
-			 * @return <code>this</code> so you can chain calls together.
-			 */
-			public Please andLimitLogTo(final long logLimit)
-			{
-				m_logLimit = logLimit;
-
-				return this;
-			}
-
-			/**
-			 * Will log the data to disk only, currently through SQLite. Data is preserved across app sessions
-			 * until (a) the user uninstalls the app, (b) the user clears the app's data, or (c) you call
-			 * one of the {@link com.idevicesinc.sweetblue.BleDevice#clearHistoricalData()} overloads.
-			 */
-			public static Please logToDisk()
-			{
-				return new Please(PersistenceLevel.DISK);
-			}
-
-			/**
-			 * Will log the data to current app memory only. When the app is destroyed the data will be lost.
-			 */
-			public static Please logToMemory()
-			{
-				return new Please(PersistenceLevel.MEMORY);
-			}
-
-			/**
-			 * Will log the data to both memory and disk - combination of {@link #logToMemory()} and {@link #logToDisk()}.
-			 */
-			public static Please logToMemoryAndDisk()
-			{
-				return new Please(PersistenceLevel.BOTH);
-			}
-
-			/**
-			 * Will not log the data.
-			 */
-			public static Please doNotLog()
-			{
-				return new Please(PersistenceLevel.NONE);
-			}
-		}
-
-		/**
-		 * Implement this method to be notified of when the library requests whether historical data should be written to a log,
-		 * and to respond with your preference of if/how this data should be written.
-		 */
-		Please onEvent(final HistoricalDataLogEvent e);
-	}
-
-	/**
-	 * Default implementation of {@link HistoricalDataLogFilter} set on {@link #historicalDataLogFilter}
-	 * that logs the most recent data reading to memory only, flushing the previous one.
-	 */
-	public static class DefaultHistoricalDataLogFilter implements HistoricalDataLogFilter
-	{
-		private static final Please DEFAULT = Please.logToMemory().andLimitLogTo(1);
-
-		@Override public Please onEvent(final HistoricalDataLogEvent e)
-		{
-			return DEFAULT;
-		}
-	}
-
-	/**
-	 * Provide an instance to {@link #historicalDataFactory} to return custom subclasses
-	 * of {@link com.idevicesinc.sweetblue.utils.HistoricalData} if you would like. For example
-	 * you might have a graphing library which requires a "Point" interface with methods <code>getX()</code>
-	 * and <code>getY()</code>. You could then create a factory that returns subclasses of
-	 * {@link com.idevicesinc.sweetblue.utils.HistoricalData} that implement this interface, so you don't
-	 * need to duplicate the data and waste memory.
-	 */
-	public static interface HistoricalDataFactory
-	{
-		/**
-		 * Return a new subclass of {@link HistoricalData} that for example implements a custom interface
-		 * for another library that handles graphing or analytics.
-		 */
-		HistoricalData newHistoricalData(final byte[] data, final EpochTime epochTime);
 	}
 	
 	/**
